+++ my-app/src/Components/Quiz/Quiz.tsx
@@ -8,7 +8,8 @@ import "./Quiz.css";
 import axios from "axios";
 import { useParams } from "react-router-dom";
 import { Nav } from "../Nav/nav";
-import Modal from "@material-ui/core/Modal";
+// import Modal from "@material-ui/core/Modal";
+import SimpleModal from "../Modal";
 
 const useTimer = makeStyles((theme: Theme) =>
   createStyles({
@@ -17,15 +18,6 @@ const useTimer = makeStyles((theme: Theme) =>
       "& > * + *": {
         marginLeft: theme.spacing(2),
       },
-    },
-
-    paper: {
-      position: "absolute",
-      width: 400,
-      backgroundColor: theme.palette.background.paper,
-      border: "2px solid #000",
-      boxShadow: theme.shadows[5],
-      padding: theme.spacing(2, 4, 3),
     },
   })
 );
@@ -37,66 +29,18 @@ const useStyles = makeStyles((theme: Theme) =>
         margin: theme.spacing(1),
       },
     },
+    paper: {
+      position: "absolute",
+      width: 400,
+      backgroundColor: theme.palette.background.paper,
+      border: "2px solid #000",
+      boxShadow: theme.shadows[5],
+      padding: theme.spacing(2, 4, 3),
+    },
   })
 );
 
 export const Quiz = () => {
-  function rand() {
-    return Math.round(Math.random() * 20) - 10;
-  }
-
-  function getModalStyle() {
-    const top = 50 + rand();
-    const left = 50 + rand();
-
-    return {
-      top: `${top}%`,
-      left: `${left}%`,
-      transform: `translate(-${top}%, -${left}%)`,
-    };
-  }
-
-  function SimpleModal() {
-    const classes = useStyles();
-
-    const [modalStyle] = React.useState(getModalStyle);
-    const [open, setOpen] = React.useState(false);
-
-    const handleOpen = () => {
-      setOpen(true);
-    };
-
-    const handleClose = () => {
-      setOpen(false);
-    };
-
-    const body = (
-      <div style={modalStyle}>
-        <h2 id="simple-modal-title">Text in a modal</h2>
-        <p id="simple-modal-description">
-          `thank you for playing!!! your score is {state.score}
-        </p>
-        <SimpleModal />
-      </div>
-    );
-
-    return (
-      <div>
-        <button type="button" onClick={handleOpen}>
-          Open Modal
-        </button>
-        <Modal
-          open={open}
-          onClose={handleClose}
-          aria-labelledby="simple-modal-title"
-          aria-describedby="simple-modal-description"
-        >
-          {body}
-        </Modal>
-      </div>
-    );
-  }
-
   const { id } = useParams();
   console.log(id);
   console.log("lien 47");
@@ -142,6 +86,7 @@ export const Quiz = () => {
   return (
     <>
       <Nav />
+      {/* <SimpleModal /> */}
       <div className="quiz-body">
         <h1 className="quiz-name">Fundamentals of Personal Finance.</h1>
         <div className="quiz-parent">
